% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_folders}
\alias{create_folders}
\title{Internal function - Creates all necessary folders in case they were not pushed to git}
\usage{
create_folders()
}
\value{
No returned value. Side effects only
}
\description{
Internal function - Creates all necessary folders in case they were not pushed to git
}
