% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_99_bird_species.r
\name{reading_99_bird_species}
\alias{reading_99_bird_species}
\title{Reading the Ecological and Functional Traits in 99 Bird Species data}
\usage{
reading_99_bird_species()
}
\value{
a list of two \code{data.tables} corresponding to both datasets: measured and
    compiled for table 1 and 2 respectively. Users can merge them if preferred.
}
\description{
Reading, after downloading if needed, the Ecological and
 Functional Traits in 99 Bird Species (reference: Renner, Swen C.; Hoesel,
 Willem van. 2017. "Ecological and Functional Traits in 99 Bird Species over
 a Large-Scale Gradient in Germany" Data 2, no. 2: 12. )
The complete archive is downloaded and stored. The authors provide 2 data sets:
    - measures they made on 28 species in 71 sites in 3 area in Germany
    (48.41 North, 9.41 East ; 51.15 North, 10.38 East ;
     52.98 North, 13.76 East)
    - measures compiled from the literature for 99 species
}
