% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_amniote_life_history_database.r
\name{reading_amniote_life_history_trait_database}
\alias{reading_amniote_life_history_trait_database}
\title{Reading the amniote life history trait data}
\usage{
reading_amniote_life_history_trait_database()
}
\value{
a \code{data.table} containing the content of the csv table
Amniote_Sparse_Table_Aug_2015.csv.
}
\description{
Reading, after downloading if needed, the Amniote life history trait data
base from the FigShare repository (https://doi.org/10.6084/m9.figshare.c.3308127.v1)
The complete archive is downloaded and stored, included meta data, figures,
references and supplementaries.
}
