% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_amphiBIO.r
\name{reading_amphiBIO}
\alias{reading_amphiBIO}
\title{Reading the amphiBIO data}
\usage{
reading_amphiBIO()
}
\value{
a \code{data.table} containing only the body mass and size traits from
the csv table AmphiBIO_v1.csv
}
\description{
Reading, after downloading if needed, the amphiBIO trait data
base from the FigShare repository (10.6084/m9.figshare.4644424.v5)
The complete archive is downloaded and stored, included metadata, references
and traits. Reference: Oliveira, Brunno Freire; Sao-Pedro, Vinicius Avelar;
Santos-Barrera, Georgina; Penone, Caterina; C. Costa, Gabriel (2017):
AmphiBIO_v1. figshare. Dataset. https://doi.org/10.6084/m9.figshare.4644424.v5
The associated article is: Oliveira, B. F. et al. AmphiBIO, a global database
for amphibian ecological traits. Sci. Data. 4:170123 doi:
https://doi.org/10.1038/sdata.2017.123 (2017).
}
\details{
Taxonomy follows Amphibian Species of the World species scientific names (Frost 2011).\tabular{lll}{
   Variable name \tab Definition \tab Unit \cr
   Body_mass_g \tab Maximum adult body mass. \tab Grams \cr
   Body_size_mm \tab Maximum adult body size. In Anura, body size is reported as snout to vent length (SVL). In Gymnophiona and Caudata, body size is reported as total length (TL). \tab Millimeter \cr
   Size_at_mature_min_mm \tab Minimum size at maturation/sexual maturity. \tab Millimeter \cr
   Size_at_mature_max_mm \tab Maximum size at maturation/sexual maturity. \tab Millimeter \cr
}
}
