% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_avian_body_sizes.r
\name{reading_avian_body_sizes}
\alias{reading_avian_body_sizes}
\title{Reading the avian body sizes data}
\usage{
reading_avian_body_sizes()
}
\value{
a \code{data.frame} containing all variables from avian_ssd_jan07.txt
}
\description{
Reading, after downloading if needed, the avian body sizes in
 the data repository: Lislevand, Terje; Figuerola, Jordi; Szekely, Tamas (2016):
 Data Paper. Data Paper. Wiley. Dataset.
 https://doi.org/10.6084/m9.figshare.3527864.v1
The associated article is: Lislevand, T., Figuerola, J. and Szekely, T. (2007),
 AVIAN BODY SIZES IN RELATION TO FECUNDITY, MATING SYSTEM, DISPLAY BEHAVIOR,
 AND RESOURCE SHARING. Ecology, 88: 1605-1605.
 https://doi.org/10.1890/06-2054
}
