% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_avian_hand_wing.r
\name{reading_avian_hwi}
\alias{reading_avian_hwi}
\title{Reading Avian Hand-wing Index Data}
\usage{
reading_avian_hwi()
}
\value{
a \code{data.frame} containing all variables from avian_ssd_jan07.txt
}
\description{
Reading, after downloading if needed, the avian hand-wing index in
 the data repository:
 Catherine Sheard. (2020). catherinesheard/Global-HWI v1.1 (v1.1). Zenodo. https://doi.org/10.5281/zenodo.3832215

The associated article is: Sheard, C., Neate-Clegg, M.H.C., Alioravainen, N. et al. Ecological drivers of global gradients in avian dispersal inferred from wing morphology. Nat Commun 11, 2463 (2020). https://doi.org/10.1038/s41467-020-16313-6
}
