% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_biotimeFish_taxonomy.r
\name{reading_biotimeFish_taxonomy}
\alias{reading_biotimeFish_taxonomy}
\title{Reading bioTIME fish section with taxonomy columns only.}
\usage{
reading_biotimeFish_taxonomy()
}
\value{
A \code{data.table} containing the list of fish species in BioTIME and
internally shared on the Google Drive. Taxa identified at the species level
only.
}
\description{
Reading bioTIME fish section with taxonomy columns only.
}
