% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_biotimePlants_taxonomy.r
\name{reading_biotimePlants_taxonomy}
\alias{reading_biotimePlants_taxonomy}
\title{Reading bioTIME PLANT section with taxonomy columns only.}
\usage{
reading_biotimePlants_taxonomy()
}
\value{
A \code{data.table} containing the list of plant species in BioTIME and
internally shared on the Google Drive. Taxa identified at the species level
only.
}
\description{
Reading bioTIME PLANT section with taxonomy columns only.
}
