% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_body_mass_of_late_quaternary_mammals.r
\name{reading_body_mass_of_late_quaternary_mammals}
\alias{reading_body_mass_of_late_quaternary_mammals}
\title{Reading the MOM data set.}
\usage{
reading_body_mass_of_late_quaternary_mammals()
}
\value{
a \code{data.table} containing MOMv3.txt content.
}
\description{
Reading, after downloading if needed, the data set: A. Smith,
Felisa; Kathleen Lyons, S.; Ernest, S. K. Morgan; E. Jones, Kate; Kaufman,
Dawn M.; Dayan, Tamar; et al. (2016): BODY MASS OF LATE QUATERNARY MAMMALS.
Wiley. Collection. https://doi.org/10.6084/m9.figshare.c.3297995.v1.
This data set is associated with the article: Smith, F.A., Lyons, S.K., Ernest,
S.K.M., Jones, K.E., Kaufman, D.M., Dayan, T., Marquet, P.A., Brown, J.H. and
Haskell, J.P. (2003), BODY MASS OF LATE QUATERNARY MAMMALS. Ecology,
84: 3403-3403. https://doi.org/10.1890/02-9003\tabular{ll}{
   Variable name \tab Variable definition \cr
   Continent \tab Continent that a species resides on. If species resides on more than one continent, a continent specific body mass is reported when available.  Thus, some mammals have multiple entries. The division between North and South America occurs at the isthmus of Panama. \cr
   Status \tab Whether species is currently present in the wild (extant); extinct as of late Pleistocene (extinct), extinct within the last 300 years (historical); or an introduction (introduction); Note these do not necessarily follow CITES or IUCN categories. \cr
   Order \tab Taxonomic order of species \cr
   Family \tab Taxonomic family of species \cr
   Genus \tab Taxonomic genus of species \cr
   Species \tab species epithet \cr
   log mass \tab Log10 transformation of Combined Mass \cr
   combined mass \tab Adult body mass averaged across males and females and geographic locations. \cr
   reference \tab Reference source for body mass information and/or status for that species; the updated electronic version of Wilson and Reeder (1993)  (6 June 2002; www.nmnh.si.edu/msw) serves as the status reference for all extant species. \cr
}


The list of references is provided in the metadata file metadata.htm.
}
