% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_carabid_traits.R
\name{reading_carabid_traits}
\alias{reading_carabid_traits}
\title{Reading carabid traits}
\usage{
reading_carabid_traits()
}
\value{
A \code{data.table} containing traits for 10 335 carabid species.
}
\description{
Reading carabid traits
}
