% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_COMBINE.r
\name{reading_COMBINE_imputed}
\alias{reading_COMBINE_imputed}
\title{Reading COMBINE imputed data}
\usage{
reading_COMBINE_imputed()
}
\value{
A \code{data.table}, for mammals
}
\description{
Reading, after downloading if needed, the trait data from:
Soria, C.D., Pacifici, M., Di Marco, M., Stephen, S.M. and Rondinini, C.
(2021), COMBINE: a coalesced mammal database of intrinsic and extrinsic
traits. Ecology. Accepted Author Manuscript e03344.
https://doi.org/10.1002/ecy.3344
}
