% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_DISPERSE.R
\name{reading_DISPERSE}
\alias{reading_DISPERSE}
\title{Reading DISPERSE data}
\usage{
reading_DISPERSE()
}
\value{
A \code{data.table}, for aquatic invertebrates
}
\description{
Reading, after downloading if needed, the trait data from:
Sarremejane, Romain; Cid, Núria; Stubbington, Rachel; Datry, Thibault;
Alp, Maria; Cañedo-Argüelles, Miguel; et al. (2020): DISPERSE – A trait
database to assess the dispersal potential of aquatic macroinvertebrates.
figshare. Collection. https://doi.org/10.6084/m9.figshare.c.5000633.v1
}
