% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/reading_ecological_traits_of_the_worlds_primates_data.r
\name{reading_ecological_traits_of_the_worlds_primates_data}
\alias{reading_ecological_traits_of_the_worlds_primates_data}
\title{Reading the Ecological traits of the world’s primates database}
\usage{
reading_ecological_traits_of_the_worlds_primates_data()
}
\value{
a \code{data.table} containing only body mass from the BodyMass_.csv table.
}
\description{
Reading, after downloading if needed, the Ecological traits of
the world’s primates. The data is stored in a Zenodo repository (reference:
Galán-Acedo, Carmen, Arroyo-Rodríguez, Víctor, Andresen, Ellen, &
Arasa-Gisbert, Ricard. (2020). Ecological traits of the world's primates
database (Version 5) [Data set]. Zenodo. http://doi.org/10.5281/zenodo.3839032
The associated article is: Galán-Acedo, C., Arroyo-Rodríguez, V., Andresen, E.
et al. Ecological traits of the world’s primates. Sci Data 6, 55 (2019).
https://doi.org/10.1038/s41597-019-0059-9
}
