% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_ELEData_mean_traits.r
\name{reading_ELEData_mean_traits}
\alias{reading_ELEData_mean_traits}
\title{Reading the ELEData AVONET data mean traits}
\usage{
reading_ELEData_mean_traits()
}
\value{
a dataframe of birds traits
}
\description{
Reading, after downloading if needed, the trait data from ELEData Avonet database
Tobias, Joseph (2021): AVONET: morphological, ecological and geographical
data for all birds (Tobias et al 2022 Ecology Letters doi:
https://doi.org/10.111/ele.13898). figshare. Dataset.
https://doi.org/10.6084/m9.figshare.16586228.v5

This database also has phylogenic data
}
