% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_elton_traits.r
\name{reading_elton_traits}
\alias{reading_elton_traits}
\title{Reading the Elton Traits data}
\usage{
reading_elton_traits()
}
\value{
A list of 2 \code{data.tables}, for mammals and birds
}
\description{
Reading, after downloading if needed, the trait data from:
Wilman, Hamish; Belmaker, Jonathan; Simpson, Jennifer; de la Rosa, Carolina;
 Rivadeneira, Marcelo M.; Jetz, Walter (2016): EltonTraits 1.0: Species-level
 foraging attributes of the world's birds and mammals. Wiley. Collection.
 https://doi.org/10.6084/m9.figshare.c.3306933.v1
The associated article is Wilman, H., Belmaker, J., Simpson, J., de la Rosa,
 C., Rivadeneira, M.M. and Jetz, W. (2014), EltonTraits 1.0: Species-level
 foraging attributes of the world's birds and mammals. Ecology, 95: 2027-2027.
 https://doi.org/10.1890/13-1917.1
}
