% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_european_amphibian_history_trait_data.r
\name{reading_european_amphibian_history_trait_data}
\alias{reading_european_amphibian_history_trait_data}
\title{Reading the European amphibians life history trait database}
\usage{
reading_european_amphibian_history_trait_data()
}
\value{
A \code{data.table} containing only biomass (in g) and length
traits (in mm).
}
\description{
Reading, after downloading if needed, the European amphibians life-
history trait database, a supplementary file associated to the article:
A database of life-history traits of European amphibians, Audrey Trochet,
Sylvain Moulherat, Olivier Calvez, Virginie M Stevens, Jean Clobert,
Dirk SSchmeller. Biodiversity Data Journal 2: e4123. doi: 10.3897/BDJ.2.e4123
}
