% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_herbivores_traits.r
\name{reading_herbivores_traits}
\alias{reading_herbivores_traits}
\title{Reading the functional trait data for late Quaternary herbivorous birds and mammals}
\usage{
reading_herbivores_traits()
}
\value{
A \code{data.table} with body mass data for numerous herbivorous species
  aggregated from the literature.
}
\description{
Reading, after downloading if needed, the trait data from:
Lundgren, Erick J; Schowanek, Simon D; Rowan, John; Middleton, Owen; Pedersen,
 Rasmus O; Wallach, Arian D; et al. (2021): Functional traits of the world's
 late Quaternary large-bodied avian and mammalian herbivores. figshare.
 Collection. https://doi.org/10.6084/m9.figshare.c.5001971.v1
The associated article is Lundgren, E.J., Schowanek, S.D., Rowan, J. et al.
Functional traits of the world's late Quaternary large-bodied avian and
mammalian herbivores. Sci Data 8, 17 (2021).
https://doi.org/10.1038/s41597-020-00788-5
}
