% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/reading_length_mass_allometries_in_amphibians_data.r
\name{reading_length_mass_allometries_in_amphibians_data}
\alias{reading_length_mass_allometries_in_amphibians_data}
\title{Reading length mass allometries in amphibians data set}
\usage{
reading_length_mass_allometries_in_amphibians_data()
}
\value{
A \code{data.table} containing biomass and SVL values aggregated from the
 literature by Santini et al.
}
\description{
Reading, after downloading if needed, the length mass allometries
 in amphibians data set. The data is available as a supplementary file with
 SANTINI, L., BENITEZ‐LOPEZ, A., FICETOLA, G.F. and HUIJBREGTS, M.A.J. (2018),
 Length-mass allometries in amphibians. Integrative Zoology, 13: 36-45.
 https://doi.org/10.1111/1749-4877.12268
}
