% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_meiri_lizards.r
\name{reading_meiri_lizards}
\alias{reading_meiri_lizards}
\title{Reading Meiri 2018 lizard traits}
\usage{
reading_meiri_lizards()
}
\value{
A \code{data.table}, for lizards
}
\description{
Reading, after downloading if needed, the trait data from:
SMeiri, Shai (2019), Data from: Traits of lizards of the world: variation
around a successful evolutionary design, Dryad, Dataset,
https://doi.org/10.5061/dryad.f6t39kj
}
