% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_rate_of_living_in_tetrapods_data.r
\name{reading_rate_of_living_in_tetrapods_data}
\alias{reading_rate_of_living_in_tetrapods_data}
\title{Reading rate of living in tetrapods data set}
\usage{
reading_rate_of_living_in_tetrapods_data()
}
\value{
A \code{data.table} containing biomass values aggregated from the
 literature by Stark et al for 4100 land vertebrates species.
}
\description{
Reading, after downloading if needed, the data in the
supplementary material associated to the article  Stark, G, Pincheira-Donoso, D,
 Meiri, S. No evidence for the "rate-of-living" theory across the tetrapod
 tree of life. Global Ecol Biogeogr. 2020; 29: 857-884.
 https://doi.org/10.1111/geb.13069
}
