% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_thermal_dependence.r
\name{reading_thermal_dependence}
\alias{reading_thermal_dependence}
\title{Reading the thermal dependence of biological traits data}
\usage{
reading_thermal_dependence()
}
\value{
A \code{data.table} with all traits. This data set does not seem to provide
   raw body size or body mass values
}
\description{
Reading, after downloading if needed, the trait data from:
Dell, Anthony, I.; Pawar, Samraat; M. Savage, Van (2016): Data Paper. Data
 Paper. Wiley. Dataset. https://doi.org/10.6084/m9.figshare.3556611.v1
The associated article is: Dell, A.I., Pawar, S. and Savage, V.M. (2013),
 The thermal dependence of biological traits. Ecology, 94: 1205-1206.
 https://doi.org/10.1890/12-2060.1
}
