% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_TRY.r
\name{reading_TRY_imputed}
\alias{reading_TRY_imputed}
\title{Reading local TRY imputed data}
\usage{
reading_TRY_imputed(path = "./data/shared_data")
}
\arguments{
\item{path}{Path to the folder where the archive is and where data will be extracted
default to ./data/shared_data/}
}
\value{
A \code{data.table}, for plants
}
\description{
Reading TRY plant trait data from the local TRY archive.
Species names are extracted from the input_data/hierarchy.info.csv table.

Trait codes are extracted from the
TRY metadata: ReadMe_TRYvs5_GapfilledData.docx.

Trait values are extracted from gapfilled_data/mean_gap_filled.txt

Back transformed data is also available in the archive.
}
