% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_TRY.r
\name{reading_TRY_raw}
\alias{reading_TRY_raw}
\title{Reading local TRY RAW data}
\usage{
reading_TRY_raw(path = "./data/shared_data")
}
\arguments{
\item{path}{Path to the folder where the archive is and where data will be extracted
default to ./data/shared_data/}
}
\value{
A \code{data.table}, for plants
}
\description{
Reading local trait data from TRY.

Species names are extracted
from the input_data/hierarchy.info.csv table.

Trait codes are extracted from the
TRY metadata: ReadMe_TRYvs5_GapfilledData.docx.

Trait values are extracted from input_data/traint.info.csv
}
