% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_UHURU_measures.r
\name{reading_UHURU_measures}
\alias{reading_UHURU_measures}
\title{Reading UHURU experimental site measure and occurrence data for mammals}
\usage{
reading_UHURU_measures()
}
\value{
A \code{data.table}, of traits for mammals
}
\description{
Reading, after downloading if needed, the trait measurements and
occurrence data from:
Alston, Jesse M.,Courtney G. Reed, Leo M. Khasoha, Bianca R.P. Brown, Gilbert
Busienei, Nathaniel Carlson, Tyler C. Coverdale, et al. 2022. “Ecological Consequences
of Large Herbivore Exclusion in an African Savanna: 12 years of Data from the UHURU
Experiment.” Ecology 103(4): e3649
}
\details{
Associated paper:
}
